Handle darwin filesystem
