+++ filesystem/filesystem_linux.go
@@ -1,47 +1,12 @@
 package filesystem
 
-import (
-	"os/exec"
-	"regexp"
-	"strings"
-)
+var dfOptions = []string{"-l"}
+var expectedLength = 6
 
-type FileSystem struct{}
-
-const name = "filesystem"
-
-func (self *FileSystem) Name() string {
-	return name
-}
-
-func (self *FileSystem) Collect() (result interface{}, err error) {
-	result, err = getFileSystemInfo()
-	return
-}
-
-func getFileSystemInfo() (fileSystemInfo map[string]interface{}, err error) {
-
-	fileSystemInfo = make(map[string]interface{})
-
-	/* Grab filesystem data from df
-	Filesystem  1K-blocks  Used  Available  Use%  Mounted on
-	*/
-
-	out, err := exec.Command("df", "-l", "-B1").Output()
-	if err != nil {
-		return
+func updatefileSystemInfo(fileSystemInfo map[string]interface{}, values []string) {
+	name := values[0]
+	fileSystemInfo[name] = map[string]string{
+		"kb_size":    values[1],
+		"mounted_on": values[5],
 	}
-	expectedLength := 6
-	lines := strings.Split(string(out), "\n")
-	for _, line := range lines[1:] {
-		values := regexp.MustCompile("\\s+").Split(line, expectedLength)
-		if len(values) == expectedLength {
-			name := values[5]
-			fileSystemInfo[name] = map[string]string{
-				"size": values[1],
-			}
-		}
-	}
-
-	return
 }
