+++ verity.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/DataDog/verity/cpu"
-	// "github.com/DataDog/verity/filesystem"
+	"github.com/DataDog/verity/filesystem"
 	// "github.com/DataDog/verity/memory"
 	// "github.com/DataDog/verity/network"
 	"github.com/DataDog/verity/platform"
@@ -19,7 +19,7 @@ type Collector interface {
 
 var collectors = []Collector{
 	&cpu.Cpu{},
-	// &filesystem.FileSystem{},
+	&filesystem.FileSystem{},
 	// &memory.Memory{},
 	// &network.Network{},
 	&platform.Platform{},
